+++ src/lj_cconv.c
@@ -199,7 +199,7 @@ void lj_cconv_ct_ct(CTState *cts, CType *d, CType *s,
     /* The conversion must exactly match the semantics of JIT-compiled code! */
     if (dsize < 8) {
       int64_t i = lj_num2i64(n);  /* Always convert via int64_t. */
-      if (dsize == 4) *(int32_t *)dp = i;
+      if (dsize == 4) *(int32_t *)dp = (int32_t)i;
       else if (dsize == 2) *(int16_t *)dp = (int16_t)i;
       else *(int8_t *)dp = (int8_t)i;
     } else if (dsize == 8) {
